Create Server View State Module
