+++ src/main/menus/app.test.js
@@ -55,7 +55,7 @@ jest.mock('common/servers/serverManager', () => ({
     getOrderedServers: jest.fn(),
     getOrderedTabsForServer: jest.fn(),
 }));
-jest.mock('main/app/servers', () => ({
+jest.mock('app/serverViewState', () => ({
     switchServer: jest.fn(),
 }));
 jest.mock('main/diagnostics', () => ({}));
