+++ src/main/menus/app.ts
@@ -6,6 +6,8 @@
 import {app, ipcMain, Menu, MenuItemConstructorOptions, MenuItem, session, shell, WebContents, clipboard} from 'electron';
 import log from 'electron-log';
 
+import ServerViewState from 'app/serverViewState';
+
 import {OPEN_SERVERS_DROPDOWN, SHOW_NEW_SERVER_MODAL} from 'common/communication';
 import {t} from 'common/utils/util';
 import {getViewDisplayName, ViewType} from 'common/views/View';
@@ -19,7 +21,6 @@ import Diagnostics from 'main/diagnostics';
 import ViewManager from 'main/views/viewManager';
 import SettingsWindow from 'main/windows/settingsWindow';
 import {selectNextView, selectPreviousView} from 'main/app/views';
-import {switchServer} from 'main/app/servers';
 
 export function createTemplate(config: Config, updateManager: UpdateManager) {
     const separatorItem: MenuItemConstructorOptions = {
@@ -266,7 +267,7 @@ export function createTemplate(config: Config, updateManager: UpdateManager) {
                 label: server.name,
                 accelerator: `${process.platform === 'darwin' ? 'Cmd+Ctrl' : 'Ctrl+Shift'}+${i + 1}`,
                 click() {
-                    switchServer(server.id);
+                    ServerViewState.switchServer(server.id);
                 },
             });
             if (ServerManager.getCurrentServer().id === server.id) {
